Make AutoRefreshModifier private and fold it into AutoRefresh.swift
